rename module and function, remove int file, put api path in config file
