+++ src/dask_flood_mapper/config.yaml
@@ -7,4 +7,5 @@
     resolution: 0.00018
     bbox: [12.3, 54.3, 13.1, 54.6]
     groupby: Null
+  api: "https://stac.eodc.eu/api/v1"
